+++ CHANGELOG.md
@@ -13,6 +13,10 @@ _Note: Gaps between patch versions are faulty, broken or test releases._
 
 ## v3.0.0-rc.??? (2021-??-??)
 
+#### :bug: Bug Fix
+
+* [Fixed an issue with tests execution took too long](https://github.com/V4Fire/Client/issues/436)
+
 #### :house: Internal
 
 * [`bVirtualScroll` is now implements `iItems` trait `base/b-virtual-scroll`](https://github.com/V4Fire/Client/issues/471)